Refactor : service class, directory, httpsessionutils
