+++ src/main/java/codesquad/controller/HomeController.java
@@ -1,7 +1,7 @@
 package codesquad.controller;
 
 
-import codesquad.domain.question.QuestionRepository;
+import codesquad.domain.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
