+++ src/main/java/codesquad/utils/HttpSessionUtils.java
@@ -1,6 +1,7 @@
 package codesquad.utils;
 
-import codesquad.domain.user.User;
+import codesquad.domain.User;
+import codesquad.exception.UserNotLoggedInException;
 
 import javax.servlet.http.HttpSession;
 
@@ -8,7 +9,7 @@ public class HttpSessionUtils {
     public static final String USER_SESSION_KEY = "sessionedUser";
 
     public static boolean isLogin(HttpSession httpSession) {
-        Object sessionedUser = httpSession.getAttribute("sessionedUser");
+        Object sessionedUser = httpSession.getAttribute(USER_SESSION_KEY);
         if (sessionedUser == null) {
             return false;
         }
@@ -17,7 +18,7 @@ public class HttpSessionUtils {
 
     public static User getSessionedUser(HttpSession httpSession) {
         if (!isLogin(httpSession)) {
-            return null;
+            throw new UserNotLoggedInException();
         }
         return (User) httpSession.getAttribute(USER_SESSION_KEY);
     }
